requirements.txt
pytesseract
pillow
pandas
openai
faiss-cpu